MiceExtVal

This package allows the user to externally validate Cox and Logistic Regression models from the data that are usually reported in the articles. The package is formed by functions that allow the user to recalibrate the models and also obtain the discrimination and calibration performance while working with multiple imputed data.

To learn more about MiceExtVal, we recommend you to start from the webpage article

https://metodologianavarrabiomed.github.io/mice-ext-val/articles/get-started.html

See also

Author

Maintainer: Asier Ballesteros Domínguez asier.ballesteros.dominguez@navarra.es (ORCID)