Calculates the predictions for a Cox model

Calculates the predictions for the given Cox model and external validation data. The predictions are calculated following the definition of a cox model

$$S_0(t)^{e^{(\beta \cdot X)}}$$

where \(S_0(t)\) stands for the survival function in a time \(t\), \(\beta\) stands for the model coefficients and \(X\) stands for the centered values in each id.

Usage

# S3 method for class 'cox'
calculate_predictions(model, data)

Arguments

model

Model generated with mv_model_cox()

data

External validation data. Multiple imputation dataset in long format.

Value

The model with the parameters predictions_aggregated, predictions_data, betax and betax_data populated.

predictions_aggregated, stores the predictions aggregated by the mean.

predictions_data, stores all the predictions in each of the imputed datasets.

betax, stores the \(\beta \cdot X\) values aggregated by the mean.

betax_data, stores the \(\beta \cdot X\) values in each of the imputed datasets.

Examples

set.seed(123)

model <- mv_model_cox(
  coefficients = list(x = 0.5, z = 0.3),
  means = list(x = 1, z = 2),
  formula = event ~ x + z,
  S0 = 0.98765
)

data <- data.frame(
  .imp = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  id = c(1, 2, 3, 1, 2, 3, 1, 2, 3),
  x = rnorm(9, 1, 0.25),
  z = rnorm(9, 2, 0.75)
)